python/graphql/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
strawberry-graphql[fastapi]==0.243.0
pydantic==2.9.2
